refactor!: move variable deletion prompting out of VariableMap.
